msm7627a: bluetooth: datatype mismatch on read bahama version
Between marimba driver and board file, there is datatype mismatch
which is causing to access a invalid memory location.

CRs-fixed: 435621

+++ arch/arm/mach-msm/lge/vee7/board-msm7627a-bt_vee7.c
@@ -495,7 +495,7 @@ static int bahama_bt(int on)
 
 	const struct bahama_config_register *p;
 
-	u8 version;
+	int version;
 
 	const struct bahama_config_register v10_bt_on[] = {
 		{ 0xE9, 0x00, 0xFF },
